requirements.txt
black==22.3.0
click==8.1.3
colorama==0.4.4
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.5.2
psycopg2-binary==2.9.3
tomli==2.0.1
